renamed main cli file
