+++ lib/cli/cli.js
